fix: unaligned pointer read during early init
fix unaligned `ptr::read_volatile` at `PMemAlloc::init` that results
panic from boot stage.
